perf(build): gateway-only compile skips Tauri/WebKit
Add a `desktop` default feature and extract the Tauri shell into desktop.rs so `cargo build --no-default-features --bin toolport-gateway` avoids WebKit and tauri_build. CI/docker-publish drops npm + WebKit apt deps; sidecar and Dockerfile.source use the same fast path.
