+++ .github/workflows/docker-publish.yml
@@ -31,17 +31,12 @@ jobs:
         with:
           persist-credentials: false
 
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: npm
-
       - uses: dtolnay/rust-toolchain@stable
 
       - name: Install Linux build dependencies
         run: |
           sudo apt-get update
-          sudo apt-get install -y libwebkit2gtk-4.1-dev libsoup-3.0-dev build-essential curl wget file libxdo-dev libssl-dev libayatana-appindicator3-dev librsvg2-dev patchelf pkg-config libdbus-1-dev
+          sudo apt-get install -y build-essential libssl-dev pkg-config
 
       - name: Cache cargo + target
         uses: actions/cache@v4
@@ -53,11 +48,11 @@ jobs:
           key: ${{ runner.os }}-cargo-${{ hashFiles('src-tauri/Cargo.lock') }}
           restore-keys: ${{ runner.os }}-cargo-
 
-      - run: npm ci
-      - run: npm run build
-
       - name: Build toolport-gateway
-        run: cargo build --release --bin toolport-gateway --manifest-path src-tauri/Cargo.toml
+        run: cargo build --release --bin toolport-gateway --manifest-path src-tauri/Cargo.toml --no-default-features
+
+      - name: Test toolport-gateway
+        run: cargo test --release --bin toolport-gateway --manifest-path src-tauri/Cargo.toml --no-default-features
 
       - uses: actions/upload-artifact@v4
         with:
