+++ Dockerfile.source
@@ -1,5 +1,5 @@
 # From-source headless Toolport gateway image (local dev when you don't have Rust).
-# Slow: compiles Tauri/WebKit inside Docker. CI uses Dockerfile (runtime-only) +
+# Uses gateway-only compile (no Tauri/WebKit). CI uses Dockerfile (runtime-only) +
 # a cached host build instead.
 #
 #   docker build -f Dockerfile.source -t toolport-gateway .
@@ -7,12 +7,10 @@
 FROM rust:1-slim-bookworm AS build
 WORKDIR /src
 RUN apt-get update && apt-get install -y --no-install-recommends \
-        pkg-config libssl-dev libdbus-1-dev libglib2.0-dev libgtk-3-dev libwebkit2gtk-4.1-dev libjavascriptcoregtk-4.1-dev \
+        pkg-config libssl-dev \
     && rm -rf /var/lib/apt/lists/*
 
-COPY src-tauri/Cargo.toml src-tauri/Cargo.lock src-tauri/build.rs src-tauri/tauri.conf.json ./src-tauri/
-COPY src-tauri/icons ./src-tauri/icons
-COPY src-tauri/capabilities ./src-tauri/capabilities
+COPY src-tauri/Cargo.toml src-tauri/Cargo.lock src-tauri/build.rs ./src-tauri/
 WORKDIR /src/src-tauri
 RUN mkdir -p src/bin && \
     printf '%s\n' \
@@ -21,11 +19,11 @@ RUN mkdir -p src/bin && \
     printf 'fn main() {}\n' > src/main.rs && \
     printf 'fn main() {}\n' > src/bin/toolport-gateway.rs && \
     printf 'fn main() {}\n' > src/bin/mock-mcp-server.rs && \
-    cargo build --release --bin toolport-gateway
+    cargo build --release --bin toolport-gateway --no-default-features
 
 COPY src-tauri/src ./src
 RUN touch src/lib.rs src/main.rs src/bin/toolport-gateway.rs src/bin/mock-mcp-server.rs && \
-    cargo build --release --bin toolport-gateway
+    cargo build --release --bin toolport-gateway --no-default-features
 
 FROM debian:bookworm-slim
 RUN apt-get update && apt-get install -y --no-install-recommends \
